#188 - move authors tests
